conversion: button-dropdown.js to tsx
